mlops/requirements.txt
# This file is auto-generated from pyproject.toml using Poetry
# To update this file, run: poetry export -f requirements.txt --output requirements.txt
# For development dependencies, run: poetry export -f requirements.txt --output requirements.txt --with dev
#
# PREFERRED: Use Poetry for dependency management
# Install Poetry: curl -sSL https://install.python-poetry.org | python3 -
# Install dependencies: poetry install
#
# FALLBACK: Use pip if Poetry is not available
# pip install -r requirements.txt

# Core data science libraries
pandas>=2.3.1,<3.0.0
numpy>=1.26.4,<2.0.0
scikit-learn>=1.7.1,<2.0.0
scipy>=1.16.1,<2.0.0
tensorflow==2.16.2

# Audio processing libraries
librosa>=0.11.0,<1.0.0
soundfile>=0.13.1,<1.0.0
pydub>=0.25.1,<1.0.0
audioread>=3.0.1,<4.0.0
soxr>=0.5.0.post1,<1.0.0
numba>=0.61.2,<1.0.0

# BirdNET integration
birdnetlib>=0.18.0,<1.0.0

keras>=3.11.1,<4.0.0

# Visualization libraries
matplotlib>=3.10.5,<4.0.0
seaborn>=0.13.2,<1.0.0
plotly>=6.2.0,<7.0.0

# ML experiment tracking
mlflow>=2.22.1,<3.0.0

# Utilities
tqdm>=4.67.1,<5.0.0
requests>=2.32.4,<3.0.0
tabulate>=0.9.0,<1.0.0
millify>=0.1.1,<1.0.0
dacite>=1.9.2,<2.0.0
PyYAML>=6.0.2,<7.0.0

# Development tools
jupyter>=1.1.1,<2.0.0
notebook>=7.4.5,<8.0.0
jupyterlab>=4.4.5,<5.0.0
ipython>=9.4.0,<10.0.0

# Filtering tools
noisereduce>=3.0.3,<4.0.0

# Graphing tools
pydot>=4.0.1,<5.0.0
graphviz>=0.21,<1.0.0

# Web app framework
streamlit>=1.48.0,<2.0.0

# Additional ML/Data utilities
joblib>=1.5.1,<2.0.0
pillow>=11.3.0,<12.0.0

#mlops requirements
mlflow[databricks]>=2.11.3
numpy>=1.23.0
pandas>=1.5.3
scikit-learn>=1.1.1
matplotlib>=3.5.2
pillow>=10.0.1
Jinja2>=3.0.3
pytz>=2022.2.1
pytest>=7.1.2
databricks-feature-engineering>=0.10.2
lightgbm>=4.6.0
